Corrects typo in 8_atoi

+++ 8_atoi/src/ft_atoi.c
@@ -32,7 +32,7 @@ int	ft_atoi(char *str)
 				if (num > INT_MAX)
 					return (INT_MIN);
 				else
-					return ((int)(sign * num));
+					return (sign * num);
 		}
 		i++;
 	}
